feat+fix(deploy): fixed deploy cmd and reworked interface
1. It's now compatible with multiple repositories
2. In addition, the interface has been changed (more explicit)

Tags and empty-commit creation logic no longer inferred. This is now
required:

If --tag is specified then a git tag will be created for the chosen git
commit.

If --empty-commit is specified then an empty commit will be made and no
git commit selection will be prompted.

If --empty-commit and --tag are specified then no prompt will be made
AND a tag will be created for the empty commit.

mim d --tag --empty-commit is the same as mim d --env production
previously

mim d --empty-commit is the same as mim d previously

However, with this explicitness, we can create empty commits in staging
(which is often very very useful).